handle novu guest user navbar
